travel/requirements.txt
aenum==2.2.3
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-filter==2.2.0
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-upload-validator==1.1.5
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.9
Jinja2==3.0.0a1
libmagic==1.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
paho-mqtt==1.5.0
Pillow==7.0.0
pycparser==2.20
pyembed==1.3.3
pymdown-extensions==7.1
python-dateutil==2.8.1
python-http-client==3.2.7
python-magic==0.4.18
python-magic-bin==0.4.14
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.42
requests==2.23.0
schedule==0.6.0
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
xlwt==1.3.0
